# api/openapi/main.yaml
openapi: 3.0.3
info:
  title: OT Continuum API
  version: 1.0.0
  description: |
    Multi-tenant operational technology risk management platform.
    
    ## Authentication
    All endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## Multi-Tenancy
    All requests are automatically scoped to the authenticated user's tenant via JWT claims.
    
    ## Rate Limiting
    - 1000 requests per hour per user
    - 10000 requests per hour per tenant
  contact:
    name: OT Continuum API Support
    email: api-support@otcontinuum.com
  license:
    name: Proprietary
    url: https://otcontinuum.com/license
servers:
  - url: https://{project_id}.supabase.co/functions/v1/make-server-fb677d93
    description: Production server
    variables:
      project_id:
        default: your-project-id
        description: Your Supabase project ID
  - url: http://localhost:54321/functions/v1/make-server-fb677d93
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /sites:
    get:
      tags: [sites]
      summary: List sites
      description: Returns sites accessible to the authenticated user
      operationId: listSites
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: criticality
          in: query
          schema:
            $ref: '#/components/schemas/Criticality'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Site'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [sites]
      summary: Create site
      description: Create a new site (admin only)
      operationId: createSite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteCreate'
      responses:
        '201':
          description: Site created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /sites/{siteId}:
    get:
      tags: [sites]
      summary: Get site by ID
      operationId: getSite
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [sites]
      summary: Update site
      description: Update site details (admin or manager)
      operationId: updateSite
      parameters:
        - $ref: '#/components/parameters/SiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteUpdate'
      responses:
        '200':
          description: Site updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [sites]
      summary: Delete site
      description: Soft delete site (admin only)
      operationId: deleteSite
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '204':
          description: Site deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # =============================================================================
  # SIGNALS
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Supabase Auth
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
...
    SiteId:
      name: siteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Criticality:
      type: string
      enum: [critical, high, medium, low]
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
    Timestamps:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # =============================================================================
    # TENANT
    # =============================================================================
...
    Site:
      allOf:
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          required:
            - id
            - tenant_id
            - name
            - code
            - criticality
            - is_active
          properties:
            id:
              type: string
              format: uuid
            tenant_id:
              type: string
              format: uuid
            name:
              type: string
            code:
              type: string
            criticality:
              $ref: '#/components/schemas/Criticality'
            site_type:
              type: string
              nullable: true
            location:
              type: object
              nullable: true
            timezone:
              type: string
            is_active:
              type: boolean
            commissioned_at:
              type: string
              format: date-time
              nullable: true
            decommissioned_at:
              type: string
              format: date-time
              nullable: true
    SiteCreate:
      type: object
      required:
        - name
        - code
        - criticality
      properties:
        name:
          type: string
        code:
          type: string
        criticality:
          $ref: '#/components/schemas/Criticality'
        site_type:
          type: string
        location:
          type: object
        timezone:
          type: string
    SiteUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        criticality:
          $ref: '#/components/schemas/Criticality'
        site_type:
          type: string
        location:
          type: object
        timezone:
          type: string
        is_active:
          type: boolean

    # =============================================================================
    # SIGNAL
    # =============================================================================
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Authentication token is required
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: forbidden
            message: You do not have permission to perform this action
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not_found
            message: The requested resource was not found
    ValidationError:
      description: Validation error - Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            error: validation_error
            message: Request validation failed
            errors:
              - field: email
                message: Invalid email format
              - field: site_id
                message: Site ID is required
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: internal_error
            message: An unexpected error occurred